adding asa calculation and rearranging the files
