+++ aiida_orca/__init__.py
@@ -1,3 +1,3 @@
 """AiiDA-Orca plugin"""
 
-__version__ = '0.3.0'
+__version__ = '0.5.0'
