+++ examples/example_1.py
@@ -11,7 +11,7 @@ from aiida.orm import load_node, Code, Dict, SinglefileData, StructureData
 from aiida.common import NotExistent
 from aiida.plugins import CalculationFactory
 
-OrcaCalculation = CalculationFactory('orca')  #pylint: disable = invalid-name
+OrcaCalculation = CalculationFactory('orca_main')  #pylint: disable = invalid-name
 
 
 def example_opt_restart(orca_code, opt_calc_pk=None, submit=True):
