+++ examples/example_2.py
@@ -10,7 +10,7 @@ from aiida.orm import (Code, Dict, StructureData)
 from aiida.common import NotExistent
 from aiida.plugins import CalculationFactory
 
-OrcaCalculation = CalculationFactory('orca')  #pylint: disable = invalid-name
+OrcaCalculation = CalculationFactory('orca_main')  #pylint: disable = invalid-name
 
 
 def example_opt_numfreq(orca_code, submit=True):
