+++ setup.json
@@ -14,7 +14,7 @@
         "License :: OSI Approved :: MIT License",
         "Framework :: AiiDA"
     ],
-    "version": "0.3.0",
+    "version": "0.5.0",
     "setup_requires": ["reentry"],
     "install_requires": [
         "aiida_core[atomic_tools] >= 1.0.0, <2.0.0",
@@ -22,7 +22,8 @@
     ],
     "entry_points": {
         "aiida.calculations": [
-            "orca = aiida_orca.calculations:OrcaCalculation"
+            "orca_main = aiida_orca.calculations:OrcaCalculation",
+            "orca_asa = aiida_orca.calculations:OrcaAsaCalculation"
         ],
         "aiida.parsers": [
             "orca_base_parser = aiida_orca.parsers:OrcaBaseParser"
